+++ client/my-sites/patterns/lib/get-pattern-permalink.ts
@@ -19,14 +19,3 @@ export function getPatternPermalink(
 	url.hash = `#pattern-${ pattern.ID }`;
 	return url.toString();
 }
-
-export const extractPatternIdFromHash = () => {
-	const pattern = /^#pattern-(\d+)$/;
-	const match = window.location.hash.match( pattern );
-
-	if ( match ) {
-		return parseInt( match[ 1 ], 10 );
-	}
-
-	return undefined;
-};
